requirements-updated.txt
# Core bot dependencies
instaloader==4.11.1
openai>=1.0.0
python-dotenv==1.0.0
schedule==1.2.0
selenium==4.15.2
python-dateutil==2.8.2
requests==2.31.0

# Auto-engagement dependencies
pandas>=1.3.0  # For data handling
numpy>=1.21.0  # For numerical operations

# Web interface dependencies
flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.4
MarkupSafe==2.1.3
